feat: OAuth2 구조 개편 시작 100% 진행
1. 카카오에서 제공한 토큰만 이용하여 유저를 특정하고 서비스 제공 성공

+++ BACKEND/rest-service/src/main/java/com/example/restservice/security/dto/UserPrincipal.java
@@ -13,7 +13,7 @@ import java.util.Collection;
 @Builder
 public class UserPrincipal implements UserDetails {
 
-    private final Integer userIndex;
+    private final Long userIndex;
     private final String userEmail;
 
     @JsonIgnore
